aska/requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3       
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3       
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0  
crispy-bootstrap4==2022.1  
Django==4.2.5
django-appconf==1.0.5      
django-crispy-forms==2.0   
django-filter==23.3        
django-select2==8.1.2      
djangorestframework==3.14.0
fontawesomefree==6.4.2     
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
mysqlclient==2.2.0
Pillow==10.0.1
PyJWT==2.8.0
python-dotenv==1.0.0       
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
twilio==8.9.0
typing_extensions==4.8.0   
tzdata==2023.3
urllib3==2.0.5
yarl==1.9.2